Eliminated duplication among the Rich* wrappers, careful not to impact performance on RichInt in particular.  Attempted to make Proxy a little bit typier.  Proxy creates an asymmetric equals method by design so it's unfixable in that regard, but achieved a minor miracle anyway by making the Rich* wrappers behave symmetrically.
Note: said miracle involved having the wrappers extend ScalaNumber in
order to induce the special == semantics.  This in turn led to implicit
conversion conflicts with the boxed types on methods like .intValue().
Resolved by moving the Rich* implicits into LowPriorityImplicits.  This
of course also removed the intentional ambiguity which prevents primitives
from automatically becoming AnyRefs.

Solved THAT one by creating dedicated, laser-precise ambiguity creating
implicits in Predef which exclude only the AnyRef methods.  Although this
is admittedly less than elegant, it is still better than it was: this way
it is direct and explicit rather than depending upon the "implicit implicit
conflict" where the barrier to promotion is intermingled with the definitions
of wrapper classes. (See the history of BoxedUnit/RichUnit for a good example
of why these concerns require separation.)

It's all worth it:

	assert(intWrapper(5) == 5)
	assert(5 == intWrapper(5))
	assert(5 == (5: java.lang.Integer))
	assert((5: java.lang.Integer) == 5)
	assert((5: java.lang.Integer) == intWrapper(5))
	assert(intWrapper(5) == (5: java.lang.Integer))

Review by community.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@23420 5e8d7ff9-d8ef-0310-90f0-a4852d11357a