Imported sbt.Process into trunk, in the guise of package scala.sys.process.  It is largely indistinguishable from the version in sbt, at least from the outside.
Also, I renamed package system to sys.  I wanted to do
that from the beginning and the desire has only grown since then.
Sometimes a short identifier is just critical to usability: with
a function like error("") called from hundreds of places, the
difference between system.error and sys.error is too big.  sys.error
and sys.exit have good vibes (at least as good as the vibes can be
for functions which error and exit.)

Note: this is just the first cut.  I need to check this in to finish
fixing partest.  I will be going over it with a comb and writing
documentation which will leave you enchanted, as well as removing
other bits which are now redundant or inferior.  No review.

git-svn-id: http://lampsvn.epfl.ch/svn-repos/scala/scala/trunk@23967 5e8d7ff9-d8ef-0310-90f0-a4852d11357a